updated for latest oF and osx sierra on 64 bit, with latest ft2xx drivers

+++ msalibs/ofxMSAControlFreakGui/src/Gui.cpp
@@ -365,7 +365,7 @@ namespace msa {
             
             //--------------------------------------------------------------
             void Gui::showControlProperties(Control *targetControl) {
-                ofLogNotice() << "Gui::showControlProperties " << (int)targetControl;
+                ofLogNotice() << "Gui::showControlProperties " << (long)targetControl;
                 _pGuiControls->_pControlProperties->show(targetControl);
             }
